@@ -14,7 +14,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(ItemEntity.class)
 public class ItemEntityMixin {
 
-    @Inject(method="Lnet/minecraft/entity/ItemEntity;isFireImmune()Z", at=@At("HEAD"), cancellable = true)
+    @Inject(method= "isFireImmune()Z", at=@At("HEAD"), cancellable = true)
     private void isFireProof(CallbackInfoReturnable<Boolean> callbackInfoReturnable) {
         ItemStack itemStack = ((ItemEntity)(Object) this).getStack();
 
